Front-end Form Validation on Login
Disables the login button if the proper requirements aren’t entered:
Added to functions that check if there is an email in the input field,
and password. Then I disabled the log in button if there was no valid
email address, no email address at all, or no password. I could’ve just
checked if the two input boxes were $pristine (unmodified) but then if
the user went and deleted what they had just inputted, the button would
work(which we don’t want). Details in the code comments.